update origin master

+++ init.sh
@@ -1,9 +1,15 @@
 (
+    if [ "$home_path" = "" ]
+    then
+        echo "[$0] Need SEOHASONG OS."
+        exit
+    fi
+
     if [ "$1" != "" ]
     then
         touch __init__.py
         touch module.py
-        cp ~/SEOHASONG/.gitignore .
+        cp $home_path/SEOHASONG/.gitignore .
         echo "# $1" > README.md
         rm -rf init.sh .git
         mv ../template.module ../$1
